Add audible byo-yomi count
This commit uses the speech synthetizer to create an audible count for remaining time in byo-yomi.
The length of the audible count is taken from "clock should start ticking at" preference. Falls back
to the tick sound if the sound synthetizer fails for some reason.

There are three styles to choose from:

- Tick (default): no change, the clock ticks in byo-yomi like it currently does in Lishogi.
- English countdown: counts in English using decreasing numbers, 1 being the last one when the
  clock starts the zeroeth second.
- Japanese count-up: starts counting upwards from 1 so that the zeroeth second is "clock should
  start ticking at" preference minus one. With a 10 second count, counts from 1 to 9 like the
  traditional timekeeper does in real life matches.